Breaking Changes for version 4.0.0 (Oculus)

+++ Archipelago.MultiClient.Net/Models/NetworkItem.cs
@@ -10,10 +10,10 @@ namespace Archipelago.MultiClient.Net.Models
     public struct NetworkItem
     {
         [JsonProperty("item")]
-        public int Item { get; set; }
+        public long Item { get; set; }
 
         [JsonProperty("location")]
-        public int Location { get; set; }
+        public long Location { get; set; }
 
         [JsonProperty("player")]
         public int Player { get; set; }
